Installs JAVA, Tomcat and Jenkins.war file
The above puppet code uses a site.pp file to override various config
settings. It downloads and installs the latest version of Java along
with Tomcat 9, it then also grabs the latest version of Jekins.war file
and publishes it. The last part creates a manager for the tomcat console